requirements.txt
fastapi
uvicorn
google-cloud-firestore
pandas
pydantic
python-multipart
pytest
httpx
